runtime/src/enclave_rpc: Move sessions
